% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapterPreamble.R
\name{chapterPreamble}
\alias{chapterPreamble}
\title{Execute chapter preamble code}
\usage{
chapterPreamble(use_cache = FALSE)
}
\arguments{
\item{use_cache}{Logical indicating whether to cache code chunks.}
}
\description{
Execute code to set up knitr compilation flags, HTML content, load required packages, etc. at the start of every chapter.
}
\details{
Compilation is performed with no tolerance for errors, no printing of package start-up messages, 
and no printing of warnings.

If the \pkg{Cairo} package is available, it will be used with the \code{"CairoPNG"} device to generate figures.

Numbers are printed to 4 digits of precision.

JS and CSS elements are set up using \code{\link{setupHTML}}.
}
\author{
Aaron Lun
}
