% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileWorkflows.R
\name{compileWorkflows}
\alias{compileWorkflows}
\title{Compile workflow chapters}
\usage{
compileWorkflows(files)
}
\arguments{
\item{files}{Character vector containing the path to individual workflow files.}
}
\value{
All files are (re)compiled to generate the corresponding \code{*_cache} directories.
\code{NULL} is invisibly returned.
}
\description{
Compile workflows for the book so that \code{\link{extractCached}} calls work correctly in other chapters.
}
\details{
Compilation is performed in an isolated session using \code{\link{r}} from the \pkg{callr} package.
This ensures that settings from one chapter do not affect the next chapter.

If an error is encountered during compilation of any Rmarkdown file,
the standard output of \code{\link{render}} leading up to the error is printed out before the function exists.
}
\author{
Aaron Lun
}
