% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettySessionInfo.R
\name{prettySessionInfo}
\alias{prettySessionInfo}
\title{Pretty session info}
\usage{
prettySessionInfo()
}
\description{
Wraps a session information markdown chunk in a collapsible element.
}
\seealso{
\code{\link{setupHTML}} and \code{\link{chapterPreamble}}, to set up the code for the collapsible element.
}
\author{
Aaron Lun
}
