% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeDependenices.R
\name{scrapeDependencies}
\alias{scrapeDependencies}
\title{Scrape dependencies}
\usage{
scrapeDependencies(dir)
}
\arguments{
\item{dir}{String specifying the directory containing Rmarkdown reports.}
}
\value{
Character vector of required packages.
}
\description{
Scrape Rmarkdown reports in the book for all required dependencies.
}
\details{
The output of this should be added to the \code{Suggests} field of \pkg{OSCAUtils}, 
to make it easier to simply install all dependencies required for the book.
}
\examples{
if (file.exists("analysis")) {
    scraped <- scrapeDependencies("analysis/")
    others <- c("GO.db")
    cat(paste0("    ", c(scraped, others)), sep=",\\n")
}
}
\author{
Aaron Lun
}
