% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupBookChapters.R
\name{setupBookChapters}
\alias{setupBookChapters}
\title{Set up the book materials}
\usage{
setupBookChapters(dir = ".")
}
\arguments{
\item{dir}{String containing a path to a Git repository in which the book is to reside.}
}
\value{
Creates or updates a Git submodule for the \pkg{OSCAbase} repository inside \code{dir}.
Creates or updates the Rmarkdown files comprising the book's contents.
Returns \code{NULL} invisibly.
}
\description{
Create the raw materials to compile the book given the files in the \pkg{OSCAbase} repository.
}
\details{
This function should be run prior to any re-compilation of the book contents,
to ensure that the latest versions of all chapters are available.

We use dynamic regeneration of book chapters in a separate repository to isolate the book source from its outputs.
This reduces the size of the \dQuote{effective} Git repository in \pkg{OSCABase},
making it easier to work with (i.e., faster cloning/pulling, no pollution of the commit history). 
Contributors are no longer burdened with the management of chapter numbering,
which if they are adding a new chapter in the middle of a part.
Our system also easily supports multiple versions of the book, e.g., for release and devel.

Note that any existing chapters will be deleted upon running this function.
Any persistent changes to chapter content should be made to the \pkg{OSCABase} repository.
}
\author{
Aaron Lun
}
