% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupHTML.R
\name{setupHTML}
\alias{setupHTML}
\title{Set up HTML elements}
\usage{
setupHTML()
}
\value{
Prints HTML to set up JS and CSS elements.
}
\description{
Set up Javascript and CSS elements for the custom collapsible class,
used in the output of \code{\link{extractCached}} and \code{\link{prettySessionInfo}}.
}
\details{
Getting this to work was a real pain in the ass.
If someone else knows better Javascript or CSS, they are more than welcome to do this in a more natural way.

It is also a bit awkward to have the same JS and CSS chunks present in every HTML document.
I guess it's not too bad, though, as it does mean that every page is standalone.
This makes things a bit easier in terms of book compilation.
}
\seealso{
\code{\link{chapterPreamble}}, which calls this function.
}
\author{
Aaron Lun
}
