% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawnBook.R
\name{spawnBook}
\alias{spawnBook}
\title{Set up the book repository}
\usage{
spawnBook(src, dir)
}
\arguments{
\item{src}{String containing a path to the directory containing the \pkg{OSCABase} contents.}

\item{dir}{String containing a path to a directory in which the compiled book is to reside.}
}
\value{
If \code{dir} does not already exist, a Git repository is created at that location.
The directory is then filled with the book contents.
A \code{NULL} is invisibly returned.
}
\description{
Initialize or update the repository containing the book contents.
}
\details{
See \code{\link{.setupBookChapters}} for some reasons why we set up the book like this.
}
\author{
Aaron Lun
}
