% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileWorkflows.R
\name{compileWorkflows}
\alias{compileWorkflows}
\title{Compile workflow chapters}
\usage{
compileWorkflows(dir = ".", fresh = TRUE, files = NULL)
}
\arguments{
\item{dir}{String containing the path to the directory of book chapters.
Usually the same as \code{dir} in \code{\link{setupBookChapters}}.}

\item{fresh}{Logical scalar indicating whether or not to delete old caches.}

\item{files}{Character vector containing the path to individual workflow files.
If this is set, \code{dir} is ignored.}
}
\value{
All workflow chapters in \code{dir} are (re)compiled to generate the corresponding \code{*_cache} directories.
(If \code{files} is specified, only those files are compiled.)
\code{NULL} is invisibly returned.
}
\description{
Compile workflows for the book so that \code{\link{extractCached}} calls work correctly in other chapters.
}
\details{
Compilation is performed in an isolated session using \code{\link{r}} from the \pkg{callr} package.
This ensures that settings from one chapter do not affect the next chapter.

The default \code{fresh=TRUE} is recommended for a rigorous recompilation,
but setting it to \code{FALSE} may be more convenient for quick debugging cycles.

If an error is encountered during compilation of any Rmarkdown file,
the standard output of \code{\link{render}} leading up to the error is printed out before the function exists.
}
\author{
Aaron Lun
}
