% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileWorkflows.R
\name{compileWorkflows}
\alias{compileWorkflows}
\title{Compile workflow chapters}
\usage{
compileWorkflows(dir = ".")
}
\arguments{
\item{dir}{String containing the path to the directory of book chapters.
Usually the same as \code{dir} in \code{\link{setupBookChapters}}.}
}
\value{
All workflow chapters are (re)compiled to generate the corresponding \code{*_cache} directories.
\code{NULL} is invisibly returned.
}
\description{
Compile workflows for the book so that \code{\link{extractCached}} calls work correctly in other chapters.
}
\details{
Compilation is performed in an isolated session using \code{\link{r}} from the \pkg{callr} package.
This ensures that settings from one chapter do not affect the next chapter.
}
\author{
Aaron Lun
}
