% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchImage.R
\name{fetchImage}
\alias{fetchImage}
\title{Fetch image}
\usage{
fetchImage(path, url)
}
\arguments{
\item{path}{String containing a path to an image file.
Can be missing or point to a file that does not exist.}

\item{url}{String containing a URL to the image file, used if \code{path} does not exist.}
}
\value{
String containing a path to an image file.
}
\description{
Fetch an image to display via \code{\link{include_graphics}}.
This either uses the path directly or obtains it from a URL if the file is not present at the indicated path.
}
\details{
This function uses \pkg{BiocFileCache} to cache the image from \code{url} if \code{path} does not point to a file.
}
\seealso{
\code{\link{include_graphics}}, to use the returned path.
}
\author{
Aaron Lun
}
